chore(libct/nsenter): extract utility code
...from nsexec.c so they can be used in cloned_binary.c.
